refactor: 헤더에서 토큰 추출하는 로직 다른 클래스로 분리
