+++ src/main/java/gift/resolver/LoginMemberArgumentResolver.java
@@ -1,5 +1,7 @@
-package gift.config;
+package gift.resolver;
 
+import gift.config.JwtProvider;
+import gift.config.LoginMember;
 import gift.exception.ErrorCode;
 import gift.exception.GiftException;
 import gift.service.MemberService;
@@ -14,10 +16,12 @@ public class LoginMemberArgumentResolver implements HandlerMethodArgumentResolve
 
     private final MemberService memberService;
     private final JwtProvider jwtProvider;
+    private final AuthHeaderManager authHeaderManager;
 
-    public LoginMemberArgumentResolver(MemberService memberService, JwtProvider jwtProvider) {
+    public LoginMemberArgumentResolver(MemberService memberService, JwtProvider jwtProvider, AuthHeaderManager authHeaderManager) {
         this.memberService = memberService;
         this.jwtProvider = jwtProvider;
+        this.authHeaderManager = authHeaderManager;
     }
 
     @Override
@@ -29,13 +33,7 @@ public class LoginMemberArgumentResolver implements HandlerMethodArgumentResolve
     public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer,
                                   NativeWebRequest webRequest, WebDataBinderFactory binderFactory) throws Exception {
         HttpServletRequest request = webRequest.getNativeRequest(HttpServletRequest.class);
-        String token = request.getHeader("Authorization");
-
-        if (token == null || !token.startsWith(jwtProvider.PREFIX)) {
-            throw new GiftException(ErrorCode.MISSING_TOKEN);
-        }
-
-        token = token.replace(jwtProvider.PREFIX, "");
+        String token = authHeaderManager.extractToken(request);
 
         if (!jwtProvider.isVerified(token)) {
             throw new GiftException(ErrorCode.INVALID_TOKEN);
